+++ java/src/test/java/com/abc/trading/events/PersistentEventStoreTest.java
@@ -1,6 +1,7 @@
 package com.abc.trading.events;
 
 import com.abc.trading.data.Quantity;
+import java.math.BigDecimal;
 import com.abc.trading.execution.LiquiditySide;
 import com.abc.trading.execution.SignalDirection;
 import com.abc.trading.msgbus.MessageBus;
@@ -45,7 +46,7 @@ class PersistentEventStoreTest {
         assertEquals(3, result.nextOffset());
         assertEquals(2, delivered.get());
         assertEquals(Quantity.fromInt(2), result.state().orders().get("order-1").filledQuantity());
-        assertEquals(2, result.state().positions().get("AAPL"));
+        assertEquals(BigDecimal.valueOf(2), result.state().positions().get("AAPL"));
         assertEquals(900.0, result.state().accounts().get("USD").free(), 1e-9);
         assertEquals(3, EventCheckpoint.load(checkpointPath).nextOffset());
     }
